requirements.txt
fastapi[all]~=0.115.5
requests
SQLAlchemy~=2.0.36
pydantic~=2.9.2
asyncpg
pydantic-settings~=2.6.1
python-jose
bcrypt==4.0.1
passlib[bcrypt]~=1.7.4
alembic~=1.14.0
pandas~=2.2.3
numpy~=2.2.0
scikit-learn~=1.6.0
redis~=5.0.0
datetime~=5.1
psycopg2-binary~=2.9.10
aiohttp~=3.9.3
elasticsearch~=8.12.1
minio~=7.2.0
aiobotocore~=2.5.0
Pillow>=10.0.0